JVMCBC-148: Fix KV response code type
Motivation
----------
In memcached protocol, response codes are defined on 2 bytes. In the
KeyValueHandler, response codes have wrongly been defined as bytes
instead of shorts, leading to bad mapping of codes into ResponseStatus
(FAILURE instead of a more precise type).

Also the fact that OPCODES are defined as byte then some received an
oversized value, casted and overflown to a single byte, is on purpose.
It should be made explicit in a comment.

Modifications
-------------
All response codes in KeyValueHandler have been migrated to shorts,
eliminating the need for casts.

Removed unnecessary casts from int/short to byte in order to make the
places where they are done on purpose stand out (KV OPCODES).

Added a test in KeyValueHandlerTest for conversion of short response
codes to ResponseStatus (conversion method has been made protected final
for that purpose).

Amended the KeyValueMessageTest integration test on getAndLock to check
that two lock attempts in a row return a TEMPORARY_FAILURE on the second
attempt, not just a FAILURE.

Results
-------
Tests allow us to better detect a mismatch in KV error interpretation.
Previously badly mapped error codes are now correctly mapped.

Reviewed-on: http://review.couchbase.org/46634
Tested-by: Simon Baslé <simon@couchbase.com>

+++ src/main/java/com/couchbase/client/core/endpoint/dcp/DCPHandler.java
@@ -62,11 +62,11 @@ import java.util.Queue;
  * @since 1.1.0
  */
 public class DCPHandler extends AbstractGenericHandler<FullBinaryMemcacheResponse, BinaryMemcacheRequest, DCPRequest> {
-    public static final byte OP_OPEN_CONNECTION = (byte) 0x50;
-    public static final byte OP_STREAM_REQUEST = (byte) 0x53;
-    public static final byte OP_SNAPSHOT_MARKER = (byte) 0x56;
-    public static final byte OP_MUTATION = (byte) 0x57;
-    public static final byte OP_REMOVE = (byte) 0x58;
+    public static final byte OP_OPEN_CONNECTION = 0x50;
+    public static final byte OP_STREAM_REQUEST = 0x53;
+    public static final byte OP_SNAPSHOT_MARKER = 0x56;
+    public static final byte OP_MUTATION = 0x57;
+    public static final byte OP_REMOVE = 0x58;
     private static final CouchbaseLogger LOGGER = CouchbaseLoggerFactory.getInstance(DCPHandler.class);
     /**
      * Maps stream identifiers to {@link DCPStream}. The identifiers put into
